# docs/swagger.yaml
basePath: /api/v1
definitions:
  middleware.ErrorJsonResponse:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: Bad Request
        type: string
    type: object
...
  presenter.CustomerJsonResponse:
    properties:
      cpf:
        example: 123.456.789-00
        type: string
      created_at:
        example: "2024-02-09T10:00:00Z"
        type: string
      email:
        example: john.doe@email.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: John Doe
        type: string
      updated_at:
        example: "2024-02-09T10:00:00Z"
        type: string
    type: object
...
  presenter.OrderJsonResponse:
    properties:
      created_at:
        example: "2024-02-09T10:00:00Z"
        type: string
      customer:
        $ref: '#/definitions/presenter.CustomerJsonResponse'
      customer_id:
        example: 1
        type: integer
      id:
        type: integer
      products:
        items:
          $ref: '#/definitions/presenter.ProductsJsonResponse'
        type: array
      status:
        example: PENDING
        type: string
      total_bill:
        example: "100.00"
        type: string
      updated_at:
        example: "2024-02-09T10:00:00Z"
        type: string
    type: object
  presenter.OrderProductJsonPaginatedResponse:
    properties:
      limit:
        example: 10
        type: integer
      order_products:
        items:
          $ref: '#/definitions/presenter.OrderProductJsonResponse'
        type: array
      page:
        example: 1
        type: integer
      total:
        example: 100
        type: integer
    type: object
  presenter.OrderProductJsonResponse:
    properties:
      created_at:
        example: "2024-02-09T10:00:00Z"
        type: string
      order:
        $ref: '#/definitions/presenter.OrderJsonResponse'
      order_id:
        type: integer
      product:
        $ref: '#/definitions/presenter.ProductJsonResponse'
      product_id:
        type: integer
      quantity:
        type: integer
      updated_at:
        example: "2024-02-09T10:00:00Z"
        type: string
    type: object
...
  presenter.ProductJsonPaginatedResponse:
    properties:
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      products:
        items:
          $ref: '#/definitions/presenter.ProductJsonResponse'
        type: array
      total:
        example: 100
        type: integer
    type: object
  presenter.ProductJsonResponse:
    properties:
      category_id:
        example: 1
        type: integer
      created_at:
        example: "2024-02-09T10:00:00Z"
        type: string
      description:
        example: Description of product A
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Product A
        type: string
      price:
        example: 99.99
        type: number
      updated_at:
        example: "2024-02-09T10:00:00Z"
        type: string
    type: object
  presenter.ProductsJsonResponse:
    properties:
      category_id:
        example: 1
        type: integer
      created_at:
        example: "2024-02-09T10:00:00Z"
        type: string
      description:
        example: Description of product A
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Product A
        type: string
      price:
        example: 99.99
        type: number
      quantity:
        type: integer
      updated_at:
        example: "2024-02-09T10:00:00Z"
        type: string
    type: object
...
  request.CreateOrderProductBodyRequest:
    properties:
      quantity:
        example: 1
        type: integer
    required:
    - quantity
    type: object
  request.CreateProductBodyRequest:
    properties:
      category_id:
        example: 1
        type: integer
      description:
        example: Product A description
        maxLength: 500
        type: string
      name:
        example: Product A
        maxLength: 100
        minLength: 3
        type: string
      price:
        example: 99.99
        type: number
    required:
    - category_id
    - name
    - price
    type: object
...
  request.UpdateOrderProductBodyRequest:
    properties:
      quantity:
        example: 1
        type: integer
    required:
    - quantity
    type: object
...
  request.UpdateProductBodyRequest:
    properties:
      category_id:
        example: 1
        type: integer
      description:
        example: Product A description
        maxLength: 500
        type: string
      name:
        example: Product A
        maxLength: 100
        minLength: 3
        type: string
      price:
        example: 99.99
        type: number
    required:
    - category_id
    - name
    - price
    type: object
...
info:
  contact: {}
  description: '### FIAP Tech Challenge Phase 3 - 10SOAT - G22'
  title: Fast Food API v3
  version: "1"
paths:
...
  /orders/products:
    get:
      consumes:
      - application/json
      description: List all order products
      parameters:
      - description: Filter by order ID
        in: query
        name: order_id
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenter.OrderProductJsonPaginatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
      summary: List order products
      tags:
      - orders
  /orders/products/{order_id}/{product_id}:
    delete:
      description: Deletes a order product by Order ID and Product ID
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: integer
      - description: Product ID
        in: path
        name: product_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenter.OrderProductJsonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
      summary: Delete order product
      tags:
      - orders
    get:
      consumes:
      - application/json
      description: Get an order product
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: integer
      - description: Product ID
        in: path
        name: product_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenter.OrderProductJsonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
      summary: Get an order product
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Create an order product
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: integer
      - description: Product ID
        in: path
        name: product_id
        required: true
        type: integer
      - description: OrderProduct data
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/request.CreateOrderProductBodyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/presenter.OrderProductJsonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
      summary: Create an order product
      tags:
      - orders
    put:
      consumes:
      - application/json
      description: Update an existing order product
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: integer
      - description: Product ID
        in: path
        name: product_id
        required: true
        type: integer
      - description: OrderProduct data
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/request.UpdateOrderProductBodyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenter.OrderProductJsonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
      summary: Update order product
      tags:
      - orders
...
  /products:
    get:
      consumes:
      - application/json
      description: |-
        List all products
        Response can return JSON or XML format (Accept header: application/json or text/xml)
      parameters:
      - description: Filter by name
        in: query
        name: name
        type: string
      - description: Filter by category ID
        in: query
        name: category_id
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      - text/xml
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenter.ProductJsonPaginatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
      summary: List products (Reference TC-1 2.b.iv)
      tags:
      - products
    post:
      consumes:
      - application/json
      description: |-
        Creates a new product
        Response can return JSON or XML format (Accept header: application/json or text/xml)
      parameters:
      - description: Product data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/request.CreateProductBodyRequest'
      produces:
      - application/json
      - text/xml
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/presenter.ProductJsonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
      summary: Create product (Reference TC-1 2.b.iii)
      tags:
      - products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: |-
        Deletes a product by ID
        Response can return JSON or XML format (Accept header: application/json or text/xml)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      - text/xml
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenter.ProductJsonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
      summary: Delete product (Reference TC-1 2.b.iii)
      tags:
      - products
    get:
      consumes:
      - application/json
      description: |-
        Search for a product by ID
        Response can return JSON or XML format (Accept header: application/json or text/xml)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      - text/xml
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenter.ProductJsonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
      summary: Get product
      tags:
      - products
    put:
      consumes:
      - application/json
      description: |-
        Update an existing product
        Response can return JSON or XML format (Accept header: application/json or text/xml)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/request.UpdateProductBodyRequest'
      produces:
      - application/json
      - text/xml
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenter.ProductJsonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
      summary: Update product (Reference TC-1 2.b.iii)
      tags:
      - products
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...